requirements.txt
bcrypt==4.0.1
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-Session==0.5.0
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
pip==23.1.2
protobuf==3.20.3
PyMySQL==1.0.2
setuptools==65.5.0
Werkzeug==2.2.3
WTForms==3.0.1
